# api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@elestyle.jp
    name: "ELESTYLE, INC"
    url: https://elepay.io
  description: elepay APIはRESTをベースに構成された決済APIです。支払い処理、返金処理など、決済に関わる運用における様々なことができます。
  title: elepay API リファレンス
  version: 1.3.0
servers:
- url: https://api.elepay.io
security:
- bearerAuth: []
- basicAuth: []
...
paths:
  /charges:
    get:
      description: 決済に関する情報を一覧で取得します。
      operationId: listCharges
      parameters:
      - description: 決済方法(複数の指定が可能です)
        explode: true
        in: query
        name: paymentMethod
        required: false
        schema:
          items:
            $ref: '#/components/schemas/PaymentMethodType'
          type: array
        style: form
      - description: 開始時間（エポックミリ秒）、指定した時間以降に作成されたデータを取得します
        explode: true
        in: query
        name: from
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: 終了時間（エポックミリ秒）、指定した時間以前に作成されたデータを取得します
        explode: true
        in: query
        name: to
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: |
          開始時間と終了時間の項目を指定します。
          - paid_time 支払い時間
          - create_time 決済新規時間
        explode: true
        in: query
        name: dateField
        required: false
        schema:
          $ref: '#/components/schemas/ChargeDateTimeType'
        style: form
      - description: 支払状況
        explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: '#/components/schemas/ChargeStatusType'
        style: form
      - description: 最大件数
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          type: integer
        style: form
      - description: 検索開始位置
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - description: |
          ソート項目
          - paid_time 支払い時間
          - create_time 決済新規時間
        explode: true
        in: query
        name: sort
        required: false
        schema:
          $ref: '#/components/schemas/ChargeDateTimeType'
        style: form
      - description: |
          ソート順
          - desc 降順
          - asc 昇順
        explode: true
        in: query
        name: order
        required: false
        schema:
          $ref: '#/components/schemas/SortOrderType'
        style: form
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ChargesResponse'
          description: OK
      summary: List charges
      tags:
      - Charge
      x-accepts: application/json;charset=utf-8
    post:
      description: 決済処理を行います。
      operationId: createCharge
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ChargeReq'
        description: 支払リクエスト
        required: true
      responses:
        "201":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ChargeDto'
          description: Created
      summary: Create charge
      tags:
      - Charge
      x-content-type: application/json;charset=utf-8
      x-accepts: application/json;charset=utf-8
  /charges/{id}:
    get:
      description: 決済に関する詳細情報を取得します。
      operationId: retrieveCharge
      parameters:
      - description: Charge ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ChargeDto'
          description: OK
      summary: Retrieve charge
      tags:
      - Charge
      x-accepts: application/json;charset=utf-8
...
components:
  schemas:
    PaymentMethodType:
      description: |
        決済方法
        - auto 自動識別(CPM決済のみ利用可能)
        - alipay アリペイ
        - alipayhk アリペイHK
        - alipayplus Alipay+
        - amazonpay Amazon Pay
        - applepay Apple Pay
        - applepay_cn Apple Pay(中国)
        - atokara アトカラ
        - atone atone(コンビニで翌月払い)
        - aupay au Pay
        - clicktopay Click To Pay
        - banktransfer 銀行振込
        - creditcard クレジットカード
        - dana DANA
        - docomopay d払い
        - ezlink EZ-Link
        - felica 電子マネー
        - felica_id iD
        - felica_quickpay QUICPay
        - felica_transport_ic 交通系ICカード
        - gcash GCash
        - ginkopay 銀行Pay
        - googlepay Google Pay
        - jcoinpay J-Coin Pay
        - jkopay JKOPAY
        - kakaopay Kakao Pay
        - konbini コンビニ決済
        - linepay LINE Pay
        - merpay メルペイ
        - origamipay Origami Pay
        - paidy Paidy 翌月払い
        - paypal PayPal
        - paypay PayPay
        - pxpayplus 全支付
        - rakutenpay 楽天ペイ
        - smartcode スマートコード
        - tng Touch 'n Go eWallet
        - truemoney TrueMoney
        - unionpay 雲閃付
        - wechatpay Wechat Pay
        - aeonpay AEON Pay
        - rabbitlinepay Rabbit LINE Pay
        - bpi BPI
        - boost Boost
        - hellomoney HelloMoney by AUB
        - tosspay Toss Pay
        - naverpay Naver Pay
        - wellwa WellWa Points
        - grabpay Grab Pay
        - momopay Momo Pay
        - promptpay Prompt Pay
        - wovenpay Woven City Pay
        - ezopay EZO Pay
      enum:
      - alipay
      - alipayhk
      - alipayplus
      - amazonpay
      - applepay
      - applepay_cn
      - atokara
      - atone
      - aupay
      - clicktopay
      - banktransfer
      - creditcard
      - dana
      - docomopay
      - ezlink
      - felica
      - felica_id
      - felica_quickpay
      - felica_transport_ic
      - gcash
      - ginkopay
      - googlepay
      - jcoinpay
      - jkopay
      - kakaopay
      - konbini
      - linepay
      - merpay
      - origamipay
      - paidy
      - paypal
      - paypay
      - pxpayplus
      - rakutenpay
      - smartcode
      - tng
      - truemoney
      - unionpay
      - wechatpay
      - aeonpay
      - rabbitlinepay
      - bpi
      - boost
      - hellomoney
      - tosspay
      - naverpay
      - wellwa
      - grabpay
      - momopay
      - promptpay
      - wovenpay
      - ezopay
      - auto
      type: string
    ResourceType:
      description: |
        決済リソース
        - web WEBブラウザ
        - ios iOS Nativeアプリ
        - android Android Nativeアプリ
        - liff Line LIFFアプリ
        - mini WeChatミニプログラム
        - cpm CPM決済、利用者提示型QRコード
        - mpm MPM決済、店舗側提示型QRコード
        - reader カードリーダー決済(リーダー設備が必要)
        - posapp POSアプリ決済(POSアプリが連携必要)
      enum:
      - web
      - ios
      - android
      - liff
      - mini
      - cpm
      - mpm
      - reader
      - posapp
      type: string
...
    ChargeDateTimeType:
      default: create_time
      description: 時間項目
      enum:
      - paid_time
      - create_time
      type: string
...
    SortOrderType:
      default: desc
      description: ソート順
      enum:
      - desc
      - asc
      type: string
    ChargeStatusType:
      description: |
        支払い状態
        - pending 未支払
        - waiting 支払い待ち(情報入力済)
        - notified 支払い準備中(通知済)
        - uncaptured 承認済み(未確定)
        - captured 支払済み
        - partially_refunded 一部返金
        - refunded 返金済み
        - amount_mismatch 金額相違
        - revoked キャンセル済み
        - failed 支払い失敗
      enum:
      - pending
      - waiting
      - notified
      - uncaptured
      - captured
      - partially_refunded
      - refunded
      - amount_mismatch
      - revoked
      - failed
      type: string
...
    ChargeReq:
      description: 支払リクエスト
      example:
        sourceId: sourceId
        amount: 0
        metadata:
          key: metadata
        orderNo: orderNo
        resource: null
        capture: true
        description: description
        locationId: locationId
        extra:
          key: extra
        clientIp: clientIp
        customerId: customerId
        paymentMethod: null
        currency: JPY
      properties:
        amount:
          description: 金額
          type: integer
        capture:
          default: true
          description: |
            支払い処理を確定するかどうか
            falseの場合、承認と支払い額の確保のみ行う。デフォルトはtrue
          type: boolean
        currency:
          default: JPY
          description: 通貨コード (ISO_4217)
          maxLength: 3
          minLength: 3
          type: string
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodType'
        resource:
          $ref: '#/components/schemas/ResourceType'
        orderNo:
          description: |
            お客様側のシステムオーダーNo（例：注文番号、決済IDなど）
            最大桁数は50桁です。
          maxLength: 50
          type: string
        description:
          description: 決済に関する説明
          maxLength: 1024
          type: string
        extra:
          additionalProperties:
            type: string
          description: 決済に関する追加情報がある場合に利用します。具体的設定情報は「開発ガイド->決済Extra情報設定」を参照してください。
          maxProperties: 20
          type: object
        metadata:
          additionalProperties:
            type: string
          description: |
            メタデータ
            キーバリューの任意データ。
            キーは20個まで、バリューは255バイト以内の文字列。
            「route」、「__」で始まる文字列は予約キー。
          maxProperties: 20
          type: object
        clientIp:
          description: Client IP アドレス
          maxLength: 39
          type: string
        customerId:
          description: カスタマID
          maxLength: 32
          type: string
        sourceId:
          description: カスタマソースID
          maxLength: 32
          type: string
        locationId:
          description: ロケーションID
          maxLength: 32
          type: string
      required:
      - amount
      - orderNo
      - paymentMethod
      type: object
    ChargeDto:
      description: 支払いオブジェクト
      example:
        metadata:
          key: metadata
        codeContent: codeContent
        description: description
        refunds:
          amount: 4
          data:
          - reason: reason
            liveMode: true
            amount: 1
            metadata:
              key: metadata
            createTime: 1
            chargeId: chargeId
            refundedTime: 1
            currency: currency
            id: id
            object: refund
            status: null
          - reason: reason
            liveMode: true
            amount: 1
            metadata:
              key: metadata
            createTime: 1
            chargeId: chargeId
            refundedTime: 1
            currency: currency
            id: id
            object: refund
            status: null
          totalCount: 7
        liveMode: true
        credential: credential
        appId: appId
        extra:
          key: extra
        currency: JPY
        refunded: true
        id: id
        authorizeAmount: 1
        authorize: true
        amount: 6
        disputed: true
        orderNo: orderNo
        cardInfo:
          last4: last4
          approvalCode: approvalCode
          wallet: wallet
          threeDSecure: true
          bankBranchName: bankBranchName
          bankAccountNo: bankAccountNo
          billingNo: billingNo
          bankName: bankName
          receivedAmount: 3
          threeDSecureVersion: threeDSecureVersion
          expMonth: 2
          repaymentType: repaymentType
          expYear: 7
          threeDSecureDetail: threeDSecureDetail
          bankUserName: bankUserName
          bankUserCode: bankUserCode
          konbiniScanTime: 9
          receivedTime: 2
          name: name
          walletApp: walletApp
          konbiniType: konbiniType
          businessType: businessType
          brand: brand
        settleTime: 4
        resource: null
        refundTime: 7
        paidTime: 6
        voucherNo: voucherNo
        createTime: 5
        clientIp: clientIp
        paid: true
        expiryTime: 1
        paymentMethod: null
        location:
          zip: zip
          note: note
          address:
            kanji:
              town: town
              city: city
              address2: address2
              address1: address1
              pref: pref
            romaji:
              town: town
              city: city
              address2: address2
              address1: address1
              pref: pref
            kana:
              town: town
              city: city
              address2: address2
              address1: address1
              pref: pref
          latitude: 5.962133916683182
          name:
            kanji: kanji
            romaji: romaji
            kana: kana
          tel: tel
          id: id
          longitude: 5.637376656633329
          status: null
        object: charge
        status: null
      properties:
        id:
          description: Charge ID
          maxLength: 32
          type: string
        appId:
          description: App ID
          maxLength: 32
          type: string
        object:
          default: charge
          description: 対象種類の表記
          type: string
        liveMode:
          description: |
            本番モードかどうか
            - false テストモード
            - true 本番モード
          type: boolean
        amount:
          description: 支払い金額
          type: integer
        authorizeAmount:
          description: 事前承認金額
          type: integer
        currency:
          default: JPY
          description: 通貨コード (ISO_4217)
          maxLength: 3
          minLength: 3
          type: string
        authorize:
          description: 事前承認フラグ
          type: boolean
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodType'
        resource:
          $ref: '#/components/schemas/ResourceType'
        orderNo:
          description: お客様システム側のオーダーNo、例えば注文番号、決済IDなど
          maxLength: 50
          type: string
        description:
          description: 支払い説明文
          maxLength: 1024
          type: string
        location:
          $ref: '#/components/schemas/ChargeLocationDto'
        extra:
          additionalProperties:
            type: string
          description: 支払いエキストラデータ
          maxProperties: 20
          type: object
        metadata:
          additionalProperties:
            type: string
          description: 支払いメタデータ
          maxProperties: 20
          type: object
        cardInfo:
          $ref: '#/components/schemas/CardInfo'
        voucherNo:
          description: プロバイダー決済伝票番号
          maxLength: 32
          type: string
        clientIp:
          description: Client IP アドレス
          maxLength: 45
          type: string
        paid:
          description: 支払い済みフラグ
          type: boolean
        refunded:
          description: 返金済みフラグ
          type: boolean
        disputed:
          description: 不審請求フラグ
          type: boolean
        refunds:
          $ref: '#/components/schemas/RefundsDto'
        status:
          $ref: '#/components/schemas/ChargeStatusType'
        codeContent:
          description: 店舗側提示型QRコード(リソースはmpm場合のみ)
          type: string
        credential:
          description: Client SDKの認証情報
          type: string
        paidTime:
          description: 支払い時間（エポックミリ秒）
          format: int64
          type: integer
        refundTime:
          description: 返金時間（エポックミリ秒）
          format: int64
          type: integer
        expiryTime:
          description: 支払い請求有効時間（エポックミリ秒）
          format: int64
          type: integer
        settleTime:
          description: 支払い締め時間（エポックミリ秒）
          format: int64
          type: integer
        createTime:
          description: 支払い新規時間（エポックミリ秒）
          format: int64
          type: integer
      type: object
...
    RefundsDto:
      description: 返金のサマリー。複数回返金された場合、返金内容をサマリーして返します。
      example:
        amount: 4
        data:
        - reason: reason
          liveMode: true
          amount: 1
          metadata:
            key: metadata
          createTime: 1
          chargeId: chargeId
          refundedTime: 1
          currency: currency
          id: id
          object: refund
          status: null
        - reason: reason
          liveMode: true
          amount: 1
          metadata:
            key: metadata
          createTime: 1
          chargeId: chargeId
          refundedTime: 1
          currency: currency
          id: id
          object: refund
          status: null
        totalCount: 7
      properties:
        amount:
          description: 返金総額
          type: integer
        totalCount:
          description: 返金回数
          type: integer
        data:
          description: 返金詳細情報
          items:
            $ref: '#/components/schemas/RefundDto'
          type: array
      type: object
    ChargesResponse:
      description: 支払い情報リスト
      example:
        total: 0
        charges:
        - metadata:
            key: metadata
          codeContent: codeContent
          description: description
          refunds:
            amount: 4
            data:
            - reason: reason
              liveMode: true
              amount: 1
              metadata:
                key: metadata
              createTime: 1
              chargeId: chargeId
              refundedTime: 1
              currency: currency
              id: id
              object: refund
              status: null
            - reason: reason
              liveMode: true
              amount: 1
              metadata:
                key: metadata
              createTime: 1
              chargeId: chargeId
              refundedTime: 1
              currency: currency
              id: id
              object: refund
              status: null
            totalCount: 7
          liveMode: true
          credential: credential
          appId: appId
          extra:
            key: extra
          currency: JPY
          refunded: true
          id: id
          authorizeAmount: 1
          authorize: true
          amount: 6
          disputed: true
          orderNo: orderNo
          cardInfo:
            last4: last4
            approvalCode: approvalCode
            wallet: wallet
            threeDSecure: true
            bankBranchName: bankBranchName
            bankAccountNo: bankAccountNo
            billingNo: billingNo
            bankName: bankName
            receivedAmount: 3
            threeDSecureVersion: threeDSecureVersion
            expMonth: 2
            repaymentType: repaymentType
            expYear: 7
            threeDSecureDetail: threeDSecureDetail
            bankUserName: bankUserName
            bankUserCode: bankUserCode
            konbiniScanTime: 9
            receivedTime: 2
            name: name
            walletApp: walletApp
            konbiniType: konbiniType
            businessType: businessType
            brand: brand
          settleTime: 4
          resource: null
          refundTime: 7
          paidTime: 6
          voucherNo: voucherNo
          createTime: 5
          clientIp: clientIp
          paid: true
          expiryTime: 1
          paymentMethod: null
          location:
            zip: zip
            note: note
            address:
              kanji:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
              romaji:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
              kana:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
            latitude: 5.962133916683182
            name:
              kanji: kanji
              romaji: romaji
              kana: kana
            tel: tel
            id: id
            longitude: 5.637376656633329
            status: null
          object: charge
          status: null
        - metadata:
            key: metadata
          codeContent: codeContent
          description: description
          refunds:
            amount: 4
            data:
            - reason: reason
              liveMode: true
              amount: 1
              metadata:
                key: metadata
              createTime: 1
              chargeId: chargeId
              refundedTime: 1
              currency: currency
              id: id
              object: refund
              status: null
            - reason: reason
              liveMode: true
              amount: 1
              metadata:
                key: metadata
              createTime: 1
              chargeId: chargeId
              refundedTime: 1
              currency: currency
              id: id
              object: refund
              status: null
            totalCount: 7
          liveMode: true
          credential: credential
          appId: appId
          extra:
            key: extra
          currency: JPY
          refunded: true
          id: id
          authorizeAmount: 1
          authorize: true
          amount: 6
          disputed: true
          orderNo: orderNo
          cardInfo:
            last4: last4
            approvalCode: approvalCode
            wallet: wallet
            threeDSecure: true
            bankBranchName: bankBranchName
            bankAccountNo: bankAccountNo
            billingNo: billingNo
            bankName: bankName
            receivedAmount: 3
            threeDSecureVersion: threeDSecureVersion
            expMonth: 2
            repaymentType: repaymentType
            expYear: 7
            threeDSecureDetail: threeDSecureDetail
            bankUserName: bankUserName
            bankUserCode: bankUserCode
            konbiniScanTime: 9
            receivedTime: 2
            name: name
            walletApp: walletApp
            konbiniType: konbiniType
            businessType: businessType
            brand: brand
          settleTime: 4
          resource: null
          refundTime: 7
          paidTime: 6
          voucherNo: voucherNo
          createTime: 5
          clientIp: clientIp
          paid: true
          expiryTime: 1
          paymentMethod: null
          location:
            zip: zip
            note: note
            address:
              kanji:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
              romaji:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
              kana:
                town: town
                city: city
                address2: address2
                address1: address1
                pref: pref
            latitude: 5.962133916683182
            name:
              kanji: kanji
              romaji: romaji
              kana: kana
            tel: tel
            id: id
            longitude: 5.637376656633329
            status: null
          object: charge
          status: null
      properties:
        total:
          description: 件数
          type: integer
        charges:
          description: 支払い詳細内容
          items:
            $ref: '#/components/schemas/ChargeDto'
          type: array
      type: object
...
    ChargeLocationStatusType:
      description: |
        ロケーションステータス
        - active 有効
        - submitted 提出済み
      enum:
      - active
      - submitted
      type: string
...
    ChargeLocationDto:
      description: ロケーションオブジェクト
      example:
        zip: zip
        note: note
        address:
          kanji:
            town: town
            city: city
            address2: address2
            address1: address1
            pref: pref
          romaji:
            town: town
            city: city
            address2: address2
            address1: address1
            pref: pref
          kana:
            town: town
            city: city
            address2: address2
            address1: address1
            pref: pref
        latitude: 5.962133916683182
        name:
          kanji: kanji
          romaji: romaji
          kana: kana
        tel: tel
        id: id
        longitude: 5.637376656633329
        status: null
      properties:
        id:
          description: Location ID
          maxLength: 32
          type: string
        name:
          $ref: '#/components/schemas/StringTransliterationDto'
        tel:
          description: 電話番号
          maxLength: 20
          type: string
        zip:
          description: 郵便番号
          maxLength: 8
          type: string
        address:
          $ref: '#/components/schemas/AddressTransliterationDto'
        latitude:
          description: 緯度
          format: double
          type: number
        longitude:
          description: 経度
          format: double
          type: number
        note:
          description: 備考
          type: string
        status:
          $ref: '#/components/schemas/ChargeLocationStatusType'
      type: object
...
    AddressTransliterationDto:
      description: 住所内容オブジェクト
      example:
        kanji:
          town: town
          city: city
          address2: address2
          address1: address1
          pref: pref
        romaji:
          town: town
          city: city
          address2: address2
          address1: address1
          pref: pref
        kana:
          town: town
          city: city
          address2: address2
          address1: address1
          pref: pref
      properties:
        kanji:
          $ref: '#/components/schemas/AddressDto'
        kana:
          $ref: '#/components/schemas/AddressDto'
        romaji:
          $ref: '#/components/schemas/AddressDto'
      type: object
    StringTransliterationDto:
      description: 名前内容オブジェクト
      example:
        kanji: kanji
        romaji: romaji
        kana: kana
      properties:
        kanji:
          description: 漢字
          type: string
        kana:
          description: カナ
          type: string
        romaji:
          description: ローマ字
          type: string
      type: object
...
  securitySchemes:
    basicAuth:
      description: |
        API呼出すの認証は、秘密鍵をユーザーネームとして扱い、パスワードは省略して、Basic 認証経由で行われます。
        秘密鍵は、全ての API リクエスト操作が可能となる重要なキーなので、取扱いにご注意ください。
      scheme: basic
      type: http
    bearerAuth:
      description: |
        API呼出すの認証は、秘密鍵を認証トークンとして扱い、Bearer 認証経由で行われます。
        秘密鍵は、全ての API リクエスト操作が可能となる重要なキーなので、取扱いにご注意ください。
      scheme: bearer
      type: http